Improve sync scheduling and document current decisions
This patch works around the fact that we're currently poorly
managing database connections, and may have SQLITE_BUSY crashes
if we interleave writes. Delays are introduced to ensure we won't
sync _right_ at the start of the application, but instead wait a
few seconds. This should give initial 'noteObservation' calls enough
time to process.

Other changes are around not doing sync work when we don't have to.